Replacing get_config handler to get_context handler in files.
