+++ calm/dsl/builtins/models/substrate.py
@@ -6,7 +6,7 @@ from .provider_spec import provider_spec
 from .client_attrs import update_dsl_metadata_map, get_dsl_metadata_map
 from .metadata_payload import get_metadata_obj
 
-from calm.dsl.config import get_config
+from calm.dsl.config import get_context
 from calm.dsl.store import Cache
 from calm.dsl.log import get_logging_handle
 
@@ -194,14 +194,14 @@ class SubstrateType(EntityType):
 
                 else:
                     # Getting the account_uuid for each provider
-                    config = get_config()
-
                     # Getting the metadata obj
                     metadata_obj = get_metadata_obj()
                     project_ref = metadata_obj.get("project_reference") or dict()
 
                     # If project not found in metadata, it will take project from config
-                    project_name = project_ref.get("name", config["PROJECT"]["name"])
+                    context_obj = get_context()
+                    project_config = context_obj.get_project_config()
+                    project_name = project_ref.get("name", project_config["name"])
 
                     project_cache_data = Cache.get_entity_data(
                         entity_type="project", name=project_name
