+++ calm/dsl/cli/mpis.py
@@ -8,7 +8,7 @@ from calm.dsl.builtins import BlueprintType, get_valid_identifier
 from calm.dsl.decompile.decompile_render import create_bp_dir
 from calm.dsl.decompile.file_handler import get_bp_dir
 from calm.dsl.api import get_api_client, get_resource_api
-from calm.dsl.config import get_config
+from calm.dsl.config import get_context
 
 from .utils import highlight_text, get_states_filter
 from .bps import launch_blueprint_simple, get_blueprint
@@ -565,9 +565,10 @@ def launch_marketplace_item(
 def convert_mpi_into_blueprint(name, version, project_name=None, app_source=None):
 
     client = get_api_client()
-    config = get_config()
+    context = get_context()
+    project_config = context.get_project_config()
 
-    project_name = project_name or config["PROJECT"]["name"]
+    project_name = project_name or project_config["name"]
     project_data = get_project(project_name)
 
     project_uuid = project_data["metadata"]["uuid"]
@@ -659,7 +660,9 @@ def publish_bp_to_marketplace_manager(
 ):
 
     client = get_api_client()
-    config = get_config()
+    context = get_context()
+    server_config = context.get_server_config()
+
     bp = get_blueprint(client, bp_name)
     bp_uuid = bp.get("metadata", {}).get("uuid", "")
 
@@ -691,7 +694,7 @@ def publish_bp_to_marketplace_manager(
             "resources": {
                 "app_attribute_list": ["FEATURED"],
                 "icon_reference_list": [],
-                "author": config["SERVER"]["pc_username"],
+                "author": server_config["pc_username"],
                 "version": version,
                 "app_group_uuid": app_group_uuid or str(uuid.uuid4()),
                 "app_blueprint_template": {
@@ -778,8 +781,9 @@ def publish_bp_as_new_marketplace_bp(
 
     if publish_to_marketplace or auto_approve:
         if not projects:
-            config = get_config()
-            projects = [config["PROJECT"]["name"]]
+            context = get_context()
+            project_config = context.get_project_config()
+            projects = [project_config["name"]]
 
         approve_marketplace_bp(
             bp_name=marketplace_bp_name,
@@ -874,8 +878,9 @@ def publish_bp_as_existing_marketplace_bp(
 
     if publish_to_marketplace or auto_approve:
         if not projects:
-            config = get_config()
-            projects = [config["PROJECT"]["name"]]
+            context = get_context()
+            project_config = context.get_project_config()
+            projects = [project_config["name"]]
 
         approve_marketplace_bp(
             bp_name=marketplace_bp_name,
