+++ calm/dsl/config/__init__.py
@@ -1,33 +1,5 @@
-from .config import (
-    get_init_data,
-    get_default_config_file,
-    get_default_db_file,
-    get_default_local_dir,
-    update_config_file_location,
-    update_init_config,
-    update_config,
-    set_config,
-    get_config,
-    print_config,
-    make_file_dir,
-)
+from .config import get_config_handle, set_dsl_config
+from .context import get_context
 
-from .context import get_context, update_config_file_context, update_project_context
-from .config2 import get_config_handle, set_dsl_config
 
-__all__ = [
-    "get_init_data",
-    "get_default_config_file",
-    "get_default_db_file",
-    "get_default_local_dir",
-    "update_config_file_location",
-    "update_init_config",
-    "update_config",
-    "set_config",
-    "get_config",
-    "print_config",
-    "make_file_dir",
-    "get_context",
-    "ConfigHandle",
-    "set_dsl_config",
-]
+__all__ = ["get_config_handle", "set_dsl_config", "get_context"]
